Adapt solution to previous commit
[Refactor] usage of int everywhere to use ulong
[Create] TextualTokenExtensions to fill in for old system
[Refactor] usage of Lexer and LexerToken to new types everywhere

+++ BakedEnv/Interpreter/Instructions/EndOfFileInstruction.cs
@@ -4,7 +4,7 @@ namespace BakedEnv.Interpreter.Instructions;
 
 public class EndOfFileInstruction : InterpreterInstruction
 {
-    public EndOfFileInstruction(int sourceIndex) : base(sourceIndex)
+    public EndOfFileInstruction(ulong sourceIndex) : base(sourceIndex)
     {
         
     }
